% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/oxygen.R
\name{calculateOxygenSolubility}
\alias{calculateOxygenSolubility}
\title{Compute oxygen solubility as a function of temperature}
\usage{
calculateOxygenSolubility(t)
}
\arguments{
\item{t}{numeric, temperature in degrees Celsius}
}
\value{
numeric vector of oxygen solubility in water values
}
\description{
This function computes the solubility of oxygen in water in freshwater
at standard pressure (760 mm). See Office of Water Quality Technical Memorandum 2011.03
(\url{https://water.usgs.gov/admin/memo/QW/qw11.03.pdf}) and
\url{http://water.usgs.gov/software/DOTABLES/}. The formula is:
\deqn{-139.34411 + \frac{1.575701e5}{t} - \frac{6.642308e7}{t^2} + \frac{1.243800e10}{t^3} - \frac{8.621949e11}{t^4}}
}
