% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/oxygen.R
\name{convertCondToSalinity}
\alias{convertCondToSalinity}
\title{Compute salinity as a function of conductivity}
\usage{
convertCondToSalinity(x)
}
\arguments{
\item{x}{conductivity}
}
\value{
a vector of salinity values
}
\description{
This function computes salinity as a function of conductivity
using equation 4 from Office of Water Quality Technical Memorandum 2011.03
\url{https://water.usgs.gov/admin/memo/QW/qw11.03.pdf}
}
