% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/oxygen.R
\name{convertDO}
\alias{convertDO}
\alias{convertDOSaturationToConcentration}
\alias{convertDOConcentrationToSaturation}
\title{Convert dissolved oxygen between \% saturation and concentration}
\usage{
convertDOSaturationToConcentration(do, t)

convertDOConcentrationToSaturation(do, t)
}
\arguments{
\item{do}{a numeric vector of dissolved oxygen measurements, \% saturation for
\code{convertDOSaturationToConcentration} and concentration in mg/L for
\code{convertDOConcentrationToSaturation}.}

\item{t}{a numeric vector of temperature measurements, in degrees Celsisus}
}
\value{
a vector of dissolved oxygen measurements, in mg/L for
\code{convertDOSaturationToConcentration} and \% saturation for
\code{convertDOConcentrationToSaturation}
}
\description{
These functions convert between dissolved oxygen measurements expressed
as \% saturation and concentration (mg / L). They use \link{calculateOxygenSolubility}
to calculate the solubility of oxygen and use that to convert.
}
\section{Functions}{
\itemize{
\item \code{convertDOSaturationToConcentration}: Convert a dissolved oxygen expressed as percent saturation to concentration.

\item \code{convertDOConcentrationToSaturation}: Convert a dissolved oxygen concentration to percent saturation.
}}

