% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculatePWQI.R
\name{createIndexFun}
\alias{createIndexFun}
\title{Creates a function to map a variable to a subindex score via a polynomial}
\usage{
createIndexFun(x)
}
\arguments{
\item{x}{a vector of polynomial coefficients}
}
\description{
Function to deal with annoying discontinuities in index; goal is
to get index function as close to continous as possible. Solve
for roots when p(x) = 10 and p(x) = 100. When there are 3 real root
choose correct real root (using deriv(p(x)) = 0 to bracket).
Use roots (correct 10 and correct 100) to clamp input to poly, use (10,100) to clamp output of poly.
Poly either has one real root so the real Filter gets it or has three real
roots and need to bracket using the derivative = 0.
See http://en.wikipedia.org/wiki/Cubic_function#The_nature_of_the_roots
}
