% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculatePWQI.R
\docType{data}
\name{poly-coef}
\alias{poly-coef}
\alias{kDORiver}
\alias{kDOStream}
\alias{kCopper}
\alias{kTSS}
\alias{kPhosphorusFanno}
\alias{kPhosphorus}
\alias{kTemperature}
\alias{kTemperatureRiver}
\alias{kMercury}
\alias{kMercuryWillamette}
\alias{kEcoli}
\alias{kAmmoniaRatio}
\alias{kDOsat}
\title{Polynomial coefficients}
\format{
An object of class \code{numeric} of length 4.

An object of class \code{numeric} of length 4.

An object of class \code{numeric} of length 4.

An object of class \code{numeric} of length 4.

An object of class \code{numeric} of length 4.

An object of class \code{numeric} of length 4.

An object of class \code{numeric} of length 4.

An object of class \code{numeric} of length 4.

An object of class \code{numeric} of length 4.

An object of class \code{numeric} of length 4.

An object of class \code{numeric} of length 4.

An object of class \code{numeric} of length 2.

An object of class \code{numeric} of length 3.
}
\usage{
kDORiver

kDOStream

kCopper

kTSS

kPhosphorusFanno

kPhosphorus

kTemperature

kTemperatureRiver

kMercury

kMercuryWillamette

kEcoli

kAmmoniaRatio

kDOsat
}
\description{
These objects are polynomial coefficients for subindex functions. There
are separate index functions for each analyte. Some analytes are split
into multiple index functions by watershed, when some watersheds have
watershed specific issues (TMDLs, etc) that need to be incorporated
into the index. The polynomial coefficients are from lower to higher
order, left to right.
}
\details{
These objects are not exported so cannot be accessed directly by users.
use \code{PWQI:::} syntax to access an object. See the examples section
for an example
}
\examples{
PWQI:::kDORiver
}
\keyword{datasets}
