% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculatePWQI.R
\name{pwqi-internal}
\alias{pwqi-internal}
\alias{Curry}
\alias{Compose}
\alias{clamp}
\alias{iclamp}
\alias{is.within}
\alias{filterRoots}
\alias{adjustDO}
\title{Internal functions}
\usage{
Curry(FUN, ...)

Compose(...)

clamp(x, a, b)

iclamp(x)

is.within(x, r)

filterRoots(p, k)

adjustDO(index, do, cel)
}
\description{
Internal functions not meant to be called directly by users.
}
