% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculatePWQI.R
\name{subindex}
\alias{subindex}
\alias{calcCu}
\alias{calcTSS}
\alias{calcDO}
\alias{calcDORiv}
\alias{calcP}
\alias{calcPF}
\alias{calcTemp}
\alias{calcTempRiv}
\alias{calcHg}
\alias{calcHgWill}
\alias{calcEcoli}
\alias{calcNH4}
\title{Subindex scoring functions}
\usage{
calcCu(...)

calcTSS(...)

calcDO(...)

calcDORiv(...)

calcP(...)

calcPF(...)

calcTemp(...)

calcTempRiv(...)

calcHg(...)

calcHgWill(...)

calcEcoli(...)

calcNH4(...)
}
\arguments{
\item{...}{a vector of raw subindex scores (e.g., copper concentrations for \code{calcCU}).}
}
\description{
These functions map subindex raw data values to subindex scores
on a scale of 10 to 100. They are all created by the \link{createIndexFun} function.
}
\examples{
calcCu(c(0,4,8))
curve(calcCu, from = 0, to = 8)
}
